Clean up application code, add multi-book and null move for users.
The new application code is much cleaner and has less clutter, with less unused
variables and simplified code.

The user can now pass on his move, so perform a null move right on the CLI. The
app now lets the user look through the history of a game after ending.