Improve endgame draw detection.
Add routine to detect KNN-K endgames and return a draw score, because there
is no forced mate.

Also add slight improvements to the opening evaluation, where central pawns
should be pushed earlier in order to open the lines for bishops. This helps
avoiding weird moves such as Bd3 and Bd6 when the d-pawn is still at d2/d7.