Add new neural network.
The format of the current neural network is Stockfish' old format. This is
due to the fact that the nnue-probe library hasn't been updated since 2020.
Stockfish' current networks have a completely different architecture (SFNNv4).

Will need to update Chess0 when a new nnue-probe library compatible with
SFNNv4 is released. I'm not planning to write a NNUE library myself.